@@ -432,139 +432,139 @@ namespace LiteDB.Queryable
 
 			All = GetMethod(
 				nameof(Queryable.All), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			AnyWithoutPredicate = GetMethod(
 				nameof(Queryable.Any), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			AnyWithPredicate = GetMethod(
 				nameof(Queryable.Any), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			AsQueryable = GetMethod(
 				nameof(Queryable.AsQueryable), 1,
-				types => new[] { typeof(IEnumerable<>).MakeGenericType(types[0]) });
+				types => [typeof(IEnumerable<>).MakeGenericType(types[0])]);
 
-			Cast = GetMethod(nameof(Queryable.Cast), 1, types => new[] { typeof(IQueryable) });
+			Cast = GetMethod(nameof(Queryable.Cast), 1, types => [typeof(IQueryable)]);
 
 			Concat = GetMethod(
 				nameof(Queryable.Concat), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0])]);
 
 			Contains = GetMethod(
 				nameof(Queryable.Contains), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), types[0] });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), types[0]]);
 
 			CountWithoutPredicate = GetMethod(
 				nameof(Queryable.Count), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			CountWithPredicate = GetMethod(
 				nameof(Queryable.Count), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			DefaultIfEmptyWithoutArgument = GetMethod(
 				nameof(Queryable.DefaultIfEmpty), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			DefaultIfEmptyWithArgument = GetMethod(
 				nameof(Queryable.DefaultIfEmpty), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), types[0] });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), types[0]]);
 
-			Distinct = GetMethod(nameof(Queryable.Distinct), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+			Distinct = GetMethod(nameof(Queryable.Distinct), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			ElementAt = GetMethod(
 				nameof(Queryable.ElementAt), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int)]);
 
 			ElementAtOrDefault = GetMethod(
 				nameof(Queryable.ElementAtOrDefault), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int)]);
 
 			Except = GetMethod(
 				nameof(Queryable.Except), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0])]);
 
 			FirstWithoutPredicate = GetMethod(
-				nameof(Queryable.First), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				nameof(Queryable.First), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			FirstWithPredicate = GetMethod(
 				nameof(Queryable.First), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			FirstOrDefaultWithoutPredicate = GetMethod(
 				nameof(Queryable.FirstOrDefault), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			FirstOrDefaultWithPredicate = GetMethod(
 				nameof(Queryable.FirstOrDefault), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			GroupByWithKeySelector = GetMethod(
 				nameof(Queryable.GroupBy), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
 			GroupByWithKeyElementSelector = GetMethod(
 				nameof(Queryable.GroupBy), 3,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1])),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[2]))
-				});
+				]);
 
 			GroupByWithKeyElementResultSelector = GetMethod(
 				nameof(Queryable.GroupBy), 4,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1])),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[2])),
 					typeof(Expression<>).MakeGenericType(
 						typeof(Func<,,>).MakeGenericType(
 							types[1], typeof(IEnumerable<>).MakeGenericType(types[2]), types[3]))
-				});
+				]);
 
 			GroupByWithKeyResultSelector = GetMethod(
 				nameof(Queryable.GroupBy), 3,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1])),
 					typeof(Expression<>).MakeGenericType(
 						typeof(Func<,,>).MakeGenericType(
 							types[1], typeof(IEnumerable<>).MakeGenericType(types[0]), types[2]))
-				});
+				]);
 
 			GroupJoin = GetMethod(
 				nameof(Queryable.GroupJoin), 4,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(IEnumerable<>).MakeGenericType(types[1]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[2])),
@@ -572,203 +572,203 @@ namespace LiteDB.Queryable
 					typeof(Expression<>).MakeGenericType(
 						typeof(Func<,,>).MakeGenericType(
 							types[0], typeof(IEnumerable<>).MakeGenericType(types[1]), types[3]))
-				});
+				]);
 
 			Intersect = GetMethod(
 				nameof(Queryable.Intersect), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0])]);
 
 			Join = GetMethod(
 				nameof(Queryable.Join), 4,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(IEnumerable<>).MakeGenericType(types[1]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[2])),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[1], types[2])),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,,>).MakeGenericType(types[0], types[1], types[3]))
-				});
+				]);
 
-			LastWithoutPredicate = GetMethod(nameof(Queryable.Last), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+			LastWithoutPredicate = GetMethod(nameof(Queryable.Last), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			LastWithPredicate = GetMethod(
 				nameof(Queryable.Last), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			LastOrDefaultWithoutPredicate = GetMethod(
 				nameof(Queryable.LastOrDefault), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			LastOrDefaultWithPredicate = GetMethod(
 				nameof(Queryable.LastOrDefault), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			LongCountWithoutPredicate = GetMethod(
 				nameof(Queryable.LongCount), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			LongCountWithPredicate = GetMethod(
 				nameof(Queryable.LongCount), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
-			MaxWithoutSelector = GetMethod(nameof(Queryable.Max), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+			MaxWithoutSelector = GetMethod(nameof(Queryable.Max), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			MaxWithSelector = GetMethod(
 				nameof(Queryable.Max), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
-			MinWithoutSelector = GetMethod(nameof(Queryable.Min), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+			MinWithoutSelector = GetMethod(nameof(Queryable.Min), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			MinWithSelector = GetMethod(
 				nameof(Queryable.Min), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
-			OfType = GetMethod(nameof(Queryable.OfType), 1, types => new[] { typeof(IQueryable) });
+			OfType = GetMethod(nameof(Queryable.OfType), 1, types => [typeof(IQueryable)]);
 
 			OrderBy = GetMethod(
 				nameof(Queryable.OrderBy), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
 			OrderByDescending = GetMethod(
 				nameof(Queryable.OrderByDescending), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
-			Reverse = GetMethod(nameof(Queryable.Reverse), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+			Reverse = GetMethod(nameof(Queryable.Reverse), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			Select = GetMethod(
 				nameof(Queryable.Select), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
 			SelectManyWithoutCollectionSelector = GetMethod(
 				nameof(Queryable.SelectMany), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(
 						typeof(Func<,>).MakeGenericType(
 							types[0], typeof(IEnumerable<>).MakeGenericType(types[1])))
-				});
+				]);
 
 			SelectManyWithCollectionSelector = GetMethod(
 				nameof(Queryable.SelectMany), 3,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(
 						typeof(Func<,>).MakeGenericType(
 							types[0], typeof(IEnumerable<>).MakeGenericType(types[1]))),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,,>).MakeGenericType(types[0], types[1], types[2]))
-				});
+				]);
 
 			SingleWithoutPredicate = GetMethod(
-				nameof(Queryable.Single), 1, types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				nameof(Queryable.Single), 1, types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			SingleWithPredicate = GetMethod(
 				nameof(Queryable.Single), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			SingleOrDefaultWithoutPredicate = GetMethod(
 				nameof(Queryable.SingleOrDefault), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0])]);
 
 			SingleOrDefaultWithPredicate = GetMethod(
 				nameof(Queryable.SingleOrDefault), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			Skip = GetMethod(
 				nameof(Queryable.Skip), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int)]);
 
 			SkipWhile = GetMethod(
 				nameof(Queryable.SkipWhile), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			Take = GetMethod(
 				nameof(Queryable.Take), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(int)]);
 
 			TakeWhile = GetMethod(
 				nameof(Queryable.TakeWhile), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
 			ThenBy = GetMethod(
 				nameof(Queryable.ThenBy), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IOrderedQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
 			ThenByDescending = GetMethod(
 				nameof(Queryable.ThenByDescending), 2,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IOrderedQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], types[1]))
-				});
+				]);
 
 			Union = GetMethod(
 				nameof(Queryable.Union), 1,
-				types => new[] { typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0]) });
+				types => [typeof(IQueryable<>).MakeGenericType(types[0]), typeof(IEnumerable<>).MakeGenericType(types[0])]);
 
 			Where = GetMethod(
 				nameof(Queryable.Where), 1,
-				types => new[]
-				{
+				types =>
+				[
 					typeof(IQueryable<>).MakeGenericType(types[0]),
 					typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], typeof(bool)))
-				});
+				]);
 
-			Type[] numericTypes = new[]
-			{
+			Type[] numericTypes =
+			[
 				typeof(int),
 				typeof(int?),
 				typeof(long),
@@ -779,7 +779,7 @@ namespace LiteDB.Queryable
 				typeof(double?),
 				typeof(decimal),
 				typeof(decimal?)
-			};
+			];
 
 			AverageWithoutSelectorMethods = new Dictionary<Type, MethodInfo>();
 			AverageWithSelectorMethods = new Dictionary<Type, MethodInfo>();
@@ -789,23 +789,23 @@ namespace LiteDB.Queryable
 			foreach(Type type in numericTypes)
 			{
 				AverageWithoutSelectorMethods[type] = GetMethod(
-					nameof(Queryable.Average), 0, types => new[] { typeof(IQueryable<>).MakeGenericType(type) });
+					nameof(Queryable.Average), 0, types => [typeof(IQueryable<>).MakeGenericType(type)]);
 				AverageWithSelectorMethods[type] = GetMethod(
 					nameof(Queryable.Average), 1,
-					types => new[]
-					{
+					types =>
+					[
 						typeof(IQueryable<>).MakeGenericType(types[0]),
 						typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], type))
-					});
+					]);
 				SumWithoutSelectorMethods[type] = GetMethod(
-					nameof(Queryable.Sum), 0, types => new[] { typeof(IQueryable<>).MakeGenericType(type) });
+					nameof(Queryable.Sum), 0, types => [typeof(IQueryable<>).MakeGenericType(type)]);
 				SumWithSelectorMethods[type] = GetMethod(
 					nameof(Queryable.Sum), 1,
-					types => new[]
-					{
+					types =>
+					[
 						typeof(IQueryable<>).MakeGenericType(types[0]),
 						typeof(Expression<>).MakeGenericType(typeof(Func<,>).MakeGenericType(types[0], type))
-					});
+					]);
 			}
 
 			MethodInfo GetMethod(string name, int genericParameterCount, Func<Type[], Type[]> parameterGenerator)
@@ -814,7 +814,7 @@ namespace LiteDB.Queryable
 					mi => ((genericParameterCount == 0 && !mi.IsGenericMethod)
 							|| (mi.IsGenericMethod && mi.GetGenericArguments().Length == genericParameterCount))
 						&& mi.GetParameters().Select(e => e.ParameterType).SequenceEqual(
-							parameterGenerator(mi.IsGenericMethod ? mi.GetGenericArguments() : Array.Empty<Type>())));
+							parameterGenerator(mi.IsGenericMethod ? mi.GetGenericArguments() : [])));
 			}
 		}
 	}
